Re-arrange JS files name and location.
 * shims.js is correctly loaded when asset pipeline is not used
 * i18n replaces i18n/translations as the main loader when asset
   pipeline is enabled. As a result, export no longer overwrites the
   loader
 * Re-enable detection of assets pipeline during export and exports to
   correct location at app/assets/javascripts/i18n
