+++ README.md
@@ -18,48 +18,55 @@ Features:
 
 ### Installation
 
-#### Rails app
+#### Rails
 
 Add the gem to your Gemfile.
 
-    source "https://rubygems.org"
-    gem "rails", "3.2.3"
+```
     gem "i18n-js"
+```
 
-#### Rails app with [Asset Pipeline](http://guides.rubyonrails.org/asset_pipeline.html)
+##### Rails with [Asset Pipeline](http://guides.rubyonrails.org/asset_pipeline.html)
 
 If you're using the [asset pipeline](http://guides.rubyonrails.org/asset_pipeline.html),
 then you must add the following line to your `app/assets/javascripts/application.js`.
 
 ```javascript
-// The following line will include `i18n` library already
-// No need to put `//= require i18n`
-//= require i18n/translations
+//= require i18n
 ```
 
-#### Rails app without [Asset Pipeline](http://guides.rubyonrails.org/asset_pipeline.html)
+##### Rails without [Asset Pipeline](http://guides.rubyonrails.org/asset_pipeline.html)
 
-If you're not using the asset pipeline, download the JavaScript file at
-<https://github.com/fnando/i18n-js/tree/master/lib/i18n.js> and load it on your page.
-Also load the `translations.js` file.
+If you're not using the asset pipeline, download the JavaScript files at
+<https://github.com/fnando/i18n-js/tree/master/app/assets/javascripts/i18n/i18n.js>,
+<https://github.com/fnando/i18n-js/tree/master/app/assets/javascripts/i18n/shims.js>
+and save it to public/javascripts.  Then load them and a file named `translations.js`
+in your application template:
 
 ```erb
 <%# This is just an example, you can put `i18n.js` and `translations.js` anywhere you like %>
-<%# Unlike the Asset Pipeline example, you need to require both **in order** %>
+<%# You need to require both **in order** %>
+<%= javascript_include_tag "shims" %>
 <%= javascript_include_tag "i18n" %>
 <%= javascript_include_tag "translations" %>
 ```
 
 This `translations.js` file can be automatically generated by the `I18n::JS::Middleware`.
 Just add it to your `config/application.rb` file.
-Don't add this middleware if you are using [asset pipeline](http://guides.rubyonrails.org/asset_pipeline.html)!
 
     config.middleware.use I18n::JS::Middleware
 
-If you can't generate this file in production (Heroku anyone?), you can "precompile"
-it by running the following command. Move the middleware line to your
-`config/environments/development.rb` file and run the following command before
-deploying.
+#### Vanilla JavaScript
+
+Just add the `i18n.js` file to your page. You'll have to build the translations object
+by hand or using your favorite programming language. See "Setting Up".
+
+### Deployment
+
+If you can't generate `translation.js` in production (Heroku anyone?), you can
+manually generate it by running the following command. Move the middleware line
+to your `config/environments/development.rb` file and run the following command
+before deploying.
 
     $ rake i18n:js:export
 
@@ -67,7 +74,7 @@ This will export all translation files, including the custom scopes you may have
 defined on `config/i18n-js.yml`. If `I18n.available_locales` is set (e.g. in your
 Rails `config/application.rb` file) then only the specified locales will be exported.
 
-#### Export Configuration
+### Export Configuration
 
 Exported translation files generated by `I18n::JS::Middleware` or `rake i18n:js:export` can be customized with config file `config/i18n-js.yml` (use `rails generate i18n:js:config` to create it). You can even get more files generated to different folders and with different translations to best suit your needs. But this does not affect anything if you use Asset Pipeline.
 
@@ -108,11 +115,6 @@ translations:
 ```
 
 To find more examples on how to use the configuration file please refer to the tests.
-
-#### Vanilla JavaScript
-
-Just add the `i18n.js` file to your page. You'll have to build the translations object
-by hand or using your favorite programming language. More info below.
 
 ### Setting up
