--- app/assets/javascripts/i18n/translations.js
@@ -1,3 +1,0 @@
-//= require i18n/shims
-//= require i18n
-//= require i18n/filtered
