+++ app/assets/javascripts/i18n/translations.js.erb
